moved rustPHastContenderRunner to RustPhastContender.cpp
